requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
idna==3.3
phonenumbers==8.12.48
pyTelegramBotAPI==4.5.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
